Add files via upload
These notebooks are just a part of work with colleagues from AUTh. Since this work is in the process of being accepted to be published I can not upload the whole text but I can give the relevant part of the abstract, to clarify at least the motivation of this work: " In this work, we examine the Swampland criteria for a specific class of rescaled $f(R)$ gravitational models, that are capable of unifying the primordial era of the Universe with the late-time era with the inclusion of string corrections. In particular, we develop separately the theoretical framework of Gauss-Bonnet and Chern-Simons theories considering that the rescale parameter is constrained to reside in the area $0<\alpha<1$. As showcased, in the context of the aforementioned theories, a viable inflationary phenomenology consistent with the latest Planck data can be obtained
for both cases for a wide variety of values. The Swampland criteria which were examined are satisfied, not necessarily simultaneously, for small values of the rescale parameter, which is in agreement with the case of a canonical scalar field with absent string corrective terms. (...) Last but not least, the power-law model which is known to be incompatible with observations is now rendered viable by including a parity violating term and as showcased, it satisfies the Swampland criteria as well.